fix: simplification de la structure des réservations
- Suppression de la relation complexe avec les catways
- Utilisation directe du catwayNumber
- Suppression des champs non nécessaires (status)
- Simplification du formatage des données
- Alignement avec la structure du fichier JSON

+++ server/models/catway.js
@@ -7,7 +7,7 @@ const catwaySchema = new mongoose.Schema({
     },
     catwayType: {
         type: String,
-        enum: ['short', 'long'],
+        enum: ['long', 'short'],
         required: true
     },
     catwayState: {
